Revert Column position to IntegerField
Fixes bug when trying to create new Columns after IDs
have passed over the 7 digit maximum (i.e. 10000).  The following
error will be reported after a ``manage reload`` operation:

InvalidOperation: quantize result has too many digits for current
context

Changing this field back to IntegerField should avoid the issue,
and the extra decimal precision isn't necessary for these objects.

+++ steelscript/appfwk/apps/datasource/models.py
@@ -663,7 +663,7 @@ class DatasourceTable(Table):
             * b/s - bits per second
             * pct - percentage
 
-        :param float position: Display position relative to other columns,
+        :param int position: Display position relative to other columns,
             automatically computed by default
 
         :param bool synthetic: Set True to compute this columns value
@@ -742,7 +742,7 @@ class Column(models.Model):
     table = models.ForeignKey(Table)
     name = models.CharField(max_length=300)
     label = models.CharField(max_length=300, null=True)
-    position = models.DecimalField(max_digits=7, decimal_places=3, default=1)
+    position = models.IntegerField(default=1)
     options = PickledObjectField()
 
     iskey = models.BooleanField(default=False)
